packages/fact-extractor/requirements.txt
docling
dspy
pandas
tqdm
python-dotenv
onnxruntime
rapidocr
easyocr
